server/requirements.txt
anyio==3.6.1
twilio
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==37.0.2
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pipenv==2022.6.7
platformdirs==2.5.2
pluggy==1.0.0
prisma==0.6.6
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
websockets==10.3

# for model image search
python-multipart
sentence_transformers
nest-asyncio
DeepImageSearch
datasets
